refactored compiler_obj.py into individual files
